+++ WalletWasabi.Fluent/Views/AddWallet/Recover/RecoverWalletView.axaml.cs
@@ -1,7 +1,7 @@
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 
-namespace WalletWasabi.Fluent.Views.AddWallet
+namespace WalletWasabi.Fluent.Views.AddWallet.Recover
 {
 	public class RecoverWalletView : UserControl
 	{
